exam_log CRUD done

+++ src/main/java/com/dragon/controller/SubjectController.java
@@ -37,7 +37,7 @@ public class SubjectController {
     }
 
     @RequestMapping("/delete")
-    public ResultSet delete(Integer id) {
+    public ResultSet delete(Long id) {
         return ResultSet.view(subjectService.deleteSubject(id));
     }
 }
